beautify this project for gods sake
