+++ apps/templates/home/course.html
@@ -193,12 +193,15 @@
             <div class="col-md-20">
                 <div class="media g-mb-30 media-comment">
                     <div class="media-body u-shadow-v18 g-bg-secondary g-pa-30">
-                        <div class="g-mb-15" style="margin-bottom: 10px">
-                            <textarea class="form-control" rows="2" placeholder="What are you thinking?"></textarea>
-                        </div>
-                        <div class="mar-top clearfix">
-                            <button type="button" class="btn btn-primary pull-right">Submit</button>
-                        </div>
+                        <form method="post">
+                            <div class="g-mb-15" style="margin-bottom: 10px">
+                                <textarea class="form-control" rows="2" placeholder="What are you thinking?"></textarea>
+                            </div>
+                            <div class="mar-top clearfix">
+                                        <input type="submit" class="btn btn-primary pull-right" value="Sumbit" >
+
+                            </div>
+                        </form>
                     </div>
                 </div>
             </div>
